refactor(navigator): split ConnectionNavigatorTree + shared categories
Extract schemaTreeCategories.ts for UnifiedSchemaTree and navigator reuse;
split ConnectionNavigatorTree into navigator/ submodules (660-line main file).
